4.21

+++ osa4/controllers/blogs.js
@@ -38,10 +38,20 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
-      await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()    // 204 No Content
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(request.params.id)
+
+        if (user.id === blog.user.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()    // 204 No Content
+        } else { 
+            return response.status(401).json({ 
+                error: 'token missing or invalid or you are not the owner of that blog' 
+            })  // 401 Unauthorized
+        }
     } catch (exception) {
-      next(exception)
+        next(exception)
     }
 })
